\docType{data}
\name{APmatrix}
\alias{APmatrix}
\title{APmatrix AP matrix of US male mortality rates, ages 0-110+, 1933-2010 (HMD)}
\description{
  APmatrix AP matrix of US male mortality rates, ages
  0-110+, 1933-2010 (HMD)
}
\references{
  \url{www.mortality.org}
}
\keyword{data}
\keyword{datasets}

