\docType{package}
\name{LexisUtils}
\alias{LexisUtils}
\alias{LexisUtils-package}
\title{A small set of utility functions for working with demographic data matrices and plotting.}
\description{
  A small set of utility functions for working with
  demographic data matrices and plotting.
}
\examples{
data(USmx)
str(USmx)
LexisMapAP(USmx)
}
\author{
  Tim Riffe \email{tim.riffe@gmail.com}
}
\references{
  \url{www.mortality.org}
}
\keyword{data}

